pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["yijun"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.13.3"
celery = "^5.3.6"
pika = "^1.3.2"
pydantic = {version = "1.10.7", extras = ["dotenv"]}
requests = "^2.31.0"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.28"
flask = "^3.0.2"
flask-pydantic = "0.11.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
